+++ lib/common_utils.dart
@@ -9,3 +9,4 @@ export 'src/log_util.dart';
 export 'src/money_util.dart';
 export 'src/timeline_util.dart';
 export 'src/endecode_util.dart';
+export 'src/text_util.dart';