+++ lib/src/date_util.dart
@@ -28,6 +28,30 @@ enum DateFormat {
   ZH_HOUR_MINUTE, //HH时mm分
 }
 
+/// 一些常用格式参照。可以自定义格式，例如："yyyy/MM/dd HH:mm:ss"，"yyyy/M/d HH:mm:ss"。
+/// 格式要求
+/// year -> yyyy/yy   month -> MM/M    day -> dd/d
+/// hour -> HH/H      minute -> mm/m   second -> ss/s
+class DataFormats {
+  static String full = "yyyy-MM-dd HH:mm:ss";
+  static String y_mo_d_h_m = "yyyy-MM-dd HH:mm";
+  static String y_mo_d = "yyyy-MM-dd";
+  static String y_mo = "yyyy-MM";
+  static String mo_d = "MM-dd";
+  static String mo_d_h_m = "MM-dd HH:mm";
+  static String h_m_s = "HH:mm:ss";
+  static String h_m = "HH:mm";
+
+  static String zh_full = "yyyy年MM月dd日 HH时mm分ss秒";
+  static String zh_y_mo_d_h_m = "yyyy年MM月dd日 HH时mm分";
+  static String zh_y_mo_d = "yyyy年MM月dd日";
+  static String zh_y_mo = "yyyy年MM月";
+  static String zh_mo_d = "MM月dd日";
+  static String zh_mo_d_h_m = "MM月dd日 HH时mm分";
+  static String zh_h_m_s = "HH时mm分ss秒";
+  static String zh_h_m = "HH时mm分";
+}
+
 /// month->days.
 Map<int, int> MONTH_DAY = {
   1: 31,
@@ -47,16 +71,23 @@ Map<int, int> MONTH_DAY = {
 /// Date Util.
 class DateUtil {
   /// get DateTime By DateStr.
-  static DateTime getDateTime(String dateStr) {
+  static DateTime getDateTime(String dateStr, {bool isUtc}) {
     DateTime dateTime = DateTime.tryParse(dateStr);
+    if (isUtc != null) {
+      if (isUtc) {
+        dateTime = dateTime.toUtc();
+      } else {
+        dateTime = dateTime.toLocal();
+      }
+    }
     return dateTime;
   }
 
   /// get DateTime By Milliseconds.
   static DateTime getDateTimeByMs(int milliseconds, {bool isUtc = false}) {
-    DateTime dateTime =
-        new DateTime.fromMillisecondsSinceEpoch(milliseconds, isUtc: isUtc);
-    return dateTime;
+    return milliseconds == null
+        ? null
+        : DateTime.fromMillisecondsSinceEpoch(milliseconds, isUtc: isUtc);
   }
 
   /// get DateMilliseconds By DateStr.
@@ -85,8 +116,9 @@ class DateUtil {
     DateFormat format = DateFormat.NORMAL,
     String dateSeparate,
     String timeSeparate,
+    bool isUtc,
   }) {
-    return getDateStrByDateTime(getDateTime(dateStr),
+    return getDateStrByDateTime(getDateTime(dateStr, isUtc: isUtc),
         format: format, dateSeparate: dateSeparate, timeSeparate: timeSeparate);
   }
 
@@ -258,6 +290,62 @@ class DateUtil {
     return time;
   }
 
+  /// format date by milliseconds.
+  /// milliseconds 日期毫秒
+  static String formatDateMs(int milliseconds,
+      {bool isUtc = false, String format}) {
+    return formatDate(getDateTimeByMs(milliseconds, isUtc: isUtc),
+        format: format);
+  }
+
+  /// format date by date str.
+  /// dateStr 日期字符串
+  static String formatDateStr(String dateStr, {bool isUtc, String format}) {
+    return formatDate(getDateTime(dateStr, isUtc: isUtc), format: format);
+  }
+
+  /// format date by DateTime.
+  /// format 转换格式(已提供常用格式 DataFormats，可以自定义格式："yyyy/MM/dd HH:mm:ss")
+  /// 格式要求
+  /// year -> yyyy/yy   month -> MM/M    day -> dd/d
+  /// hour -> HH/H      minute -> mm/m   second -> ss/s
+  static String formatDate(DateTime dateTime, {bool isUtc, String format}) {
+    if (dateTime == null) return "";
+    format = format ?? DataFormats.full;
+    if (format.contains("yy")) {
+      String year = dateTime.year.toString();
+      if (format.contains("yyyy")) {
+        format = format.replaceAll("yyyy", year);
+      } else {
+        format = format.replaceAll(
+            "yy", year.substring(year.length - 2, year.length));
+      }
+    }
+
+    format = _comFormat(dateTime.month, format, 'M', 'MM');
+    format = _comFormat(dateTime.day, format, 'd', 'dd');
+    format = _comFormat(dateTime.hour, format, 'H', 'HH');
+    format = _comFormat(dateTime.minute, format, 'm', 'mm');
+    format = _comFormat(dateTime.second, format, 's', 'ss');
+    format = _comFormat(dateTime.millisecond, format, 'S', 'SSS');
+
+    return format;
+  }
+
+  /// com format.
+  static String _comFormat(
+      int value, String format, String single, String full) {
+    if (format.contains(single)) {
+      if (format.contains(full)) {
+        format =
+            format.replaceAll(full, value < 10 ? '0$value' : value.toString());
+      } else {
+        format = format.replaceAll(single, value.toString());
+      }
+    }
+    return format;
+  }
+
   /// get WeekDay By Milliseconds.
   static String getWeekDayByMs(int milliseconds, {bool isUtc = false}) {
     DateTime dateTime = getDateTimeByMs(milliseconds, isUtc: isUtc);
@@ -355,8 +443,7 @@ class DateUtil {
   /// 是否是昨天.
   static bool isYesterday(DateTime dateTime, DateTime locDateTime) {
     if (yearIsEqual(dateTime, locDateTime)) {
-      int spDay =
-          DateUtil.getDayOfYear(locDateTime) - DateUtil.getDayOfYear(dateTime);
+      int spDay = getDayOfYear(locDateTime) - getDayOfYear(dateTime);
       return spDay == 1;
     } else {
       return ((locDateTime.year - dateTime.year == 1) &&
@@ -369,8 +456,9 @@ class DateUtil {
 
   /// get day of year.
   /// 在今年的第几天.
-  static int getDayOfYearByMillis(int millis) {
-    return getDayOfYear(DateTime.fromMillisecondsSinceEpoch(millis));
+  static int getDayOfYearByMillis(int millis, {bool isUtc = false}) {
+    return getDayOfYear(
+        DateTime.fromMillisecondsSinceEpoch(millis, isUtc: isUtc));
   }
 
   /// get day of year.
@@ -401,13 +489,31 @@ class DateUtil {
     return dateTime.year == locDateTime.year;
   }
 
-  /// year is today.
-  /// 是否是今天.
+  /// is today.
+  /// 是否是当天.
   static bool isToday(int milliseconds, {bool isUtc = false}) {
     if (milliseconds == null || milliseconds == 0) return false;
     DateTime old =
         DateTime.fromMillisecondsSinceEpoch(milliseconds, isUtc: isUtc);
     DateTime now = isUtc ? DateTime.now().toUtc() : DateTime.now().toLocal();
     return old.year == now.year && old.month == now.month && old.day == now.day;
+  }
+
+  /// is Week.
+  /// 是否是本周.
+  static bool isWeek(int milliseconds, {bool isUtc = false}) {
+    if (milliseconds == null || milliseconds <= 0) {
+      return false;
+    }
+    DateTime _old =
+        DateTime.fromMillisecondsSinceEpoch(milliseconds, isUtc: isUtc);
+    DateTime _now = isUtc ? DateTime.now().toUtc() : DateTime.now().toLocal();
+    DateTime old =
+        _now.millisecondsSinceEpoch > _old.millisecondsSinceEpoch ? _old : _now;
+    DateTime now =
+        _now.millisecondsSinceEpoch > _old.millisecondsSinceEpoch ? _now : _old;
+    return (now.weekday >= old.weekday) &&
+        (now.millisecondsSinceEpoch - old.millisecondsSinceEpoch <=
+            7 * 24 * 60 * 60 * 1000);
   }
 }