+++ lib/src/text_util.dart
@@ -1,0 +1,53 @@
+/**
+ * @Author: Sky24n
+ * @GitHub: https://github.com/Sky24n
+ * @Description: Text Util.
+ * @Date: 2019/7/9
+ */
+class TextUtil {
+  /// isEmpty
+  static bool isEmpty(String text) {
+    return text == null || text.isEmpty;
+  }
+
+  /// 每隔 x位 加 pattern
+  static String formatDigitPattern(String text,
+      {int digit = 4, String pattern = ' '}) {
+    text = text?.replaceAllMapped(new RegExp("(.{$digit})"), (Match match) {
+      return "${match.group(0)}$pattern";
+    });
+    if (text != null && text.endsWith(pattern)) {
+      text = text.substring(0, text.length - 1);
+    }
+    return text;
+  }
+
+  /// 每隔4位加空格
+  static String formatSpace4(String text) {
+    return formatDigitPattern(text);
+  }
+
+  /// 每隔3三位加逗号
+  /// num 数字或数字字符串。int型。
+  static String formatComma3(Object num) {
+    return formatDigitPattern(num?.toString(), digit: 3, pattern: ',');
+  }
+
+  /// hidePhone
+  static String hidePhone(String phoneNo,
+      {int start = 3, int end = 7, String replacement = '****'}) {
+    return phoneNo?.replaceRange(start, end, replacement);
+  }
+
+  /// replace
+  static String replace(String text, Pattern from, String replace) {
+    return text?.replaceAll(from, replace);
+  }
+
+  /// split
+  static List<String> split(String text, Pattern pattern,
+      {List<String> defValue: const []}) {
+    List<String> list = text?.split(pattern);
+    return list ?? defValue;
+  }
+}